openvino: split quantized weight extraction
